Making datatypes 64-bit safe, mainly by renaming long datatypes to UInt32. Removing semicolon from the Tdecode OS X macros.
git-svn-id: svn://svn.code.sf.net/p/playerpro/code/branches/PlayerPRO6@176 f4605116-7794-4d0b-9e48-65ca9bfeedfd

+++ MADDriver.source/MADDriver.Source/Import-Export/AMF.h
@@ -50,10 +50,10 @@ typedef struct {
     uchar       type;
     char        name[32],filename[13];
     void        *sample;
-    ulong       size;
+    UInt32      size;
     ushort      rate;
     uchar       volume;
-    ulong       loopstart,loopend;
+    UInt32      loopstart,loopend;
 } INSTRUMENT;
 
 #if PRAGMA_STRUCT_ALIGN
